requirements.txt
beautifulsoup4==4.14.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
curl_cffi==0.13.0
frozendict==2.4.6
idna==3.11
multitasking==0.0.12
numpy==2.3.4
pandas==2.3.3
peewee==3.18.3
platformdirs==4.5.0
protobuf==6.33.0
pycparser==2.23
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
six==1.17.0
soupsieve==2.8
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
yfinance==0.2.66

# Additional packages used by ETL scripts
praw==7.9.0
python-dotenv==1.0.0
vaderSentiment==3.3.3
sec-api==1.0.32
tqdm==4.67.1
pyarrow==12.0.0
arxiv==2.0.0
